fix: move dnsPacketBytes to dns-helpers and document
