+++ apps/ensindexer/src/handlers/Registry.ts
@@ -3,17 +3,21 @@ import schema from "ponder:schema";
 import { encodeLabelhash } from "@ensdomains/ensjs/utils";
 import { type Address, zeroAddress } from "viem";
 
+import {
+  type LabelHash,
+  type Node,
+  PluginName,
+  REVERSE_ROOT_NODES,
+  isLabelIndexable,
+  makeSubdomainNode,
+  maybeHealLabelByReverseAddress,
+} from "@ensnode/utils";
+
 import { makeSharedEventValues, upsertAccount, upsertResolver } from "@/lib/db-helpers";
 import { labelByLabelHash } from "@/lib/graphnode-helpers";
 import { makeResolverId } from "@/lib/ids";
 import { type EventWithArgs, healReverseAddresses } from "@/lib/ponder-helpers";
 import { recursivelyRemoveEmptyDomainFromParentSubdomainCount } from "@/lib/subgraph-helpers";
-import { type LabelHash, type Node, PluginName, REVERSE_ROOT_NODES } from "@ensnode/utils";
-import {
-  isLabelIndexable,
-  makeSubdomainNode,
-  maybeHealLabelByReverseAddress,
-} from "@ensnode/utils/subname-helpers";
 
 /**
  * makes a set of shared handlers for a Registry contract
