+++ apps/ensindexer/src/plugins/lineanames/handlers/Registrar.ts
@@ -1,7 +1,6 @@
 import { ponder } from "ponder:registry";
 
-import { type LabelHash, PluginName } from "@ensnode/utils";
-import { uint256ToHex32 } from "@ensnode/utils/subname-helpers";
+import { type LabelHash, PluginName, uint256ToHex32 } from "@ensnode/utils";
 
 import { makeRegistrarHandlers } from "@/handlers/Registrar";
 import { ENSIndexerPluginHandlerArgs } from "@/lib/plugin-helpers";
