+++ packages/ensnode-utils/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./constants";
 export * from "./types";
 export * from "./cache";
+export * from "./subname-helpers";
